erosion with points implemented. seaLevel changed to float. world elevation changed to be in [0, 3]

+++ core/src/stonering/generators/worldgen/generators/drainage/OceanFiller.java
@@ -13,7 +13,7 @@ public class OceanFiller extends AbstractGenerator {
 	private Random random;
 	private int width;
 	private int height;
-	private int seaLevel;
+	private float seaLevel;
 
 	public OceanFiller(WorldGenContainer container) {
 		super(container);
